Move more files to cpp
