+++ libgadget/forcetree.cpp
