+++ libgadget/powerspectrum.cpp
